+++ lib/core/features/city_detail_page/city_detail_page.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
-import 'package:local_turism/commons/app_strings.dart';
-import 'package:local_turism/data/models/city_model.dart';
-import 'package:local_turism/style/app_colors.dart';
-import 'package:local_turism/style/style.dart';
-import 'package:local_turism/views/pages/city_detail_page/widgets/city_fact_widget.dart';
+import 'package:local_turism/core/commons/app_strings.dart';
+import 'package:local_turism/core/models/city_model.dart';
+import 'package:local_turism/core/commons/app_colors.dart';
+import 'package:local_turism/core/commons/style.dart';
+import 'package:local_turism/core/features/city_detail_page/widgets/city_fact_widget.dart';
 import 'package:local_turism/views/widgets/photo_widget.dart';
 import 'package:local_turism/views/widgets/temperature_widget.dart';
