+++ lib/core/features/city_detail_page/widgets/city_fact_widget.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:local_turism/data/models/city_model.dart';
-import 'package:local_turism/style/app_colors.dart';
-import 'package:local_turism/style/style.dart';
+import 'package:local_turism/core/models/city_model.dart';
+import 'package:local_turism/core/commons/app_colors.dart';
+import 'package:local_turism/core/commons/style.dart';
 import 'package:local_turism/views/widgets/photo_widget.dart';
 
 class CityFactWidget extends StatelessWidget {
