+++ lib/core/models/city_model.dart
