+++ lib/core/models/weather_model.dart
